+++ code-snapshots/13-forms/16-multi-input-validator/src/app/auth/login-driver/login-driver.component.css
@@ -1,0 +1,11 @@
+.control-row {
+    align-items: flex-end;
+  }
+  
+  .control {
+    flex: 1;
+  }
+  
+  button {
+    padding: 0.6rem 1.75rem;
+  }
