--- code-snapshots/13-forms/16-multi-input-validator/src/app/auth/login/login.component.html
@@ -1,25 +1,0 @@
-<form [formGroup]="form" (ngSubmit)="onSubmit()">
-  <h2>Login</h2>
-
-  <div class="control-row">
-    <div class="control no-margin">
-      <label for="email">Email</label>
-      <input id="email" type="email" [formControl]="form.controls.email" />
-    </div>
-
-    <div class="control no-margin">
-      <label for="password">Password</label>
-      <input id="password" type="password" formControlName="password" />
-    </div>
-
-    <button class="button">Login</button>
-  </div>
-
-  @if (emailIsInvalid) {
-    <p class="control-error">Please enter a valid email address.</p>
-  }
-
-  @if (passwordIsInvalid) {
-    <p class="control-error">Please enter a valid password (must be at least 6 characters long and contain a question mark).</p>
-  }
-</form>
